pyproject.toml
[project]
name = "skills-service"
version = "0.1.0"
description = "Standalone embedding microservice for crew agent skills"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-huggingface>=0.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
